Massive refactoring and design improvements.
- All class libraries have been switched to be static and to use the CRTP pattern.
  For this application, due to the high amounts of __declspec (naked)
  custom asm functions, it is much better in design and maintenance if
  all classes are static rather than instanced. This allows us to easily
  and cleanly place all of our custom code in their respective
  libraries, while allowing us to point to them using function pointers
  within our naked function assembly.

- LibraryUtility has been renamed to LibraryLoader and is also now a
static class.

- VersionUtility is now the only location in the application that contains
the game version. The rest of the code will use the VersionUtility
functions in order to retrieve the version (so that they can do whatever
they need to do including as the basis for offset retrieval in
Library.h) or to check to see if the Game version is supported by the
mod. Each library no longer contains its own copy of the Version.

- Removed resolution x and y macros