Moved all 1.09B function redirects to their respective functions.
- All functions in D2Functions.cpp that were being "overridden" and
wrapped around after the initial retrieval of that function have been
now moved into their respective functions. The corresponding function
implementation for the specific D2 version that is being ran will
automatically be retrieved when the user first requests the function.
This means that since this is all encapsulated, we can simply request
the function we want, and everything else is taken care of in the
background. The next step is to move the remaining 1.13 specific
function redirections into their functions as well and any remaining
loose functions.

- The dll has had some compiler and linker configuration changes in
order to make it more stable. Some of these changes and recommendation
of compiler configuration have been noted in the README.