Access violations related to loading custom files at launch should now be fixed
The custom file loading function (Which is responsible for loading
LocalizedStrings.ini) has been completely rewritten. It no longer uses the
"InstallPath" directory which was causing an access violation for some users
during startup. This was because Diablo II automatically sets the InstallPath
during Diablo II Launch, and also every time you join/exit a game. The install
path location in some situations doesn't seem to be correct and thus
causes Alpaca to fail to find its own Localization file. With this change,
from my experience with working with this code base, there are no more registry
keys that are being used, with the exception of the "Save Path" key. This is needed
since we need to know where the save directory is so that we can place the shared
stash there and any other character data manipulation stuff. If there
are any errors related to registry keys or if anyone scans the code and
finds any usages of registry keys other than the "Save Path" key, please
let me know.